pyproject.toml
[tool.poetry]
name = "ugines_bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.5.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.5.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
